requirements.txt
streamlit==1.31.1
groq==0.4.2
edge-tts==6.1.9
python-dotenv==1.0.0
audio-recorder-streamlit==0.0.5
requests==2.31.0 